requirements.txt
absl-py==1.2.0
antlr4-python3-runtime==4.9.3
black==22.3.0
brotlipy==0.7.0
cachetools==5.2.0
certifi==2022.9.14
cffi @ file:///tmp/abs_98z5h56wf8/croots/recipe/cffi_1659598650955/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.3
cloudpickle==2.1.0
cryptography @ file:///tmp/build/80754af9/cryptography_1652101588893/work
cycler==0.11.0
# Editable install with no version control (data-compression==0.2)
-e /data1/fengry/projects/dtmV
# Editable install with no version control (detectron2==0.6)
-e /data1/fengry/projects/DCM_crosscheck/DCM/packages/detectron2
einops==0.4.1
fairscale==0.4.8
fonttools==4.37.1
future==0.18.2
fvcore==0.1.5.post20220512
google-auth==2.11.0
google-auth-oauthlib==0.4.6
grpcio==1.48.1
hydra-core==1.2.0
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
importlib-metadata==4.12.0
iopath==0.1.9
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186066731/work
mkl-service==2.4.0
mypy-extensions==0.4.3
ninja==1.10.2.3
numpy @ file:///tmp/abs_653_j00fmm/croots/recipe/numpy_and_numpy_base_1659432701727/work
oauthlib==3.2.0
omegaconf==2.2.3
opencv-python==4.6.0.66
packaging==21.3
pathspec==0.10.1
Pillow==9.2.0
platformdirs==2.5.2
portalocker==2.5.1
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.4
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydot==1.4.2
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.0.9
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
python-dateutil==2.8.2
PyYAML==6.0
requests @ file:///opt/conda/conda-bld/requests_1657734628632/work
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.9.1
six @ file:///tmp/build/80754af9/six_1644875935023/work
tabulate==0.8.10
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
timm==0.6.7
tomli==2.0.1
torch==1.12.1
torchac==0.9.3
torchaudio==0.12.1
torchvision==0.13.1
tqdm==4.64.1
typing_extensions @ file:///tmp/abs_ben9emwtky/croots/recipe/typing_extensions_1659638822008/work
urllib3 @ file:///tmp/abs_5dhwnz6atv/croots/recipe/urllib3_1659110457909/work
Werkzeug==2.2.2
yacs==0.1.8
zipp==3.8.1
